+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.12-SNAPSHOT] - 2022-05-23
+## [0.1.13-SNAPSHOT] - 2022-05-27
 
 ### Added
  - added property and functionality to skip protocol validation
+
+## [0.1.12-SNAPSHOT] - 2022-05-19
+ 
+### Changed
+ - New version of Multipart Message Library
+ - Modified self description document to use dateTimeStamp instead dateTime, compatibility with standard 
 
 ## [0.1.11-SNAPSHOT] - 2022-02-21